Fix invariant violations: provider-agnostic types, rename safety, retry resilience, error logging
- Move AssetItemType, AssetVersionSize, ChangeReason from icloud/photos/types
  to core types.rs so the download module has no iCloud dependency (Inv. 3)
- Extract rename_part_to_final() helper that attempts rename first, handles
  concurrent duplicate gracefully by cleaning up redundant .part (Inv. 6)
- Increase deferred state write retries from 3 to 6 with exponential backoff
  (200ms base) to survive transient SQLite contention (Inv. 6)
- Replace silent meta.modified().ok() with explicit match that logs a warning
  and treats the file as stale, forcing re-download (Inv. 7)
